+++ src/Block/Element/AbstractBlock.php
@@ -20,6 +20,8 @@ use League\CommonMark\Node\Node;
 
 /**
  * Block-level element
+ *
+ * @method parent() ?self
  */
 abstract class AbstractBlock extends Node
 {
@@ -200,7 +202,10 @@ abstract class AbstractBlock extends Node
         $this->open = false;
         $this->endLine = $endLineNumber;
 
-        $context->setTip($context->getTip()->parent());
+        // This should almost always be true
+        if ($context->getTip() !== null) {
+            $context->setTip($context->getTip()->parent());
+        }
     }
 
     /**