+++ src/DocParser.php
@@ -15,17 +15,17 @@
 namespace League\CommonMark;
 
 use League\CommonMark\Block\Element\AbstractBlock;
+use League\CommonMark\Block\Element\AbstractStringContainerBlock;
 use League\CommonMark\Block\Element\Document;
-use League\CommonMark\Block\Element\InlineContainerInterface;
 use League\CommonMark\Block\Element\Paragraph;
 use League\CommonMark\Block\Element\StringContainerInterface;
 
-class DocParser
+final class DocParser implements DocParserInterface
 {
     /**
      * @var EnvironmentInterface
      */
-    protected $environment;
+    private $environment;
 
     /**
      * @var InlineParserEngine
@@ -49,9 +49,13 @@ class DocParser
 
     /**
      * @return EnvironmentInterface
+     *
+     * @deprecated
      */
     public function getEnvironment(): EnvironmentInterface
     {
+        @trigger_error('DocParser::getEnvironment() has been deprecated and will be removed in a future release.', E_USER_DEPRECATED);
+
         return $this->environment;
     }
 
@@ -81,7 +85,7 @@ class DocParser
      */
     public function parse(string $input): Document
     {
-        $context = new Context(new Document(), $this->getEnvironment());
+        $context = new Context(new Document(), $this->environment);
         $context->setEncoding(\mb_detect_encoding($input, 'ASCII,UTF-8', true) ?: 'ISO-8859-1');
 
         $lines = $this->preProcessInput($input);
@@ -141,7 +145,7 @@ class DocParser
 
     private function processDocument(ContextInterface $context)
     {
-        foreach ($this->getEnvironment()->getDocumentProcessors() as $documentProcessor) {
+        foreach ($this->environment->getDocumentProcessors() as $documentProcessor) {
             $documentProcessor->processDocument($context->getDocument());
         }
     }
@@ -156,7 +160,7 @@ class DocParser
             }
 
             $node = $event->getNode();
-            if ($node instanceof InlineContainerInterface) {
+            if ($node instanceof AbstractStringContainerBlock) {
                 $this->inlineParserEngine->parse($node, $context->getDocument()->getReferenceMap());
             }
         }
@@ -208,7 +212,7 @@ class DocParser
                 }
             }
 
-            if (!$parsed || $context->getContainer() instanceof StringContainerInterface || $context->getTip()->getDepth() >= $this->maxNestingLevel) {
+            if (!$parsed || $context->getContainer() instanceof StringContainerInterface || (($tip = $context->getTip()) && $tip->getDepth() >= $this->maxNestingLevel)) {
                 $context->setBlocksParsed(true);
                 break;
             }