+++ src/Inline/Element/Strong.php
@@ -14,6 +14,13 @@
 
 namespace League\CommonMark\Inline\Element;
 
-class Strong extends AbstractInlineContainer
+class Strong extends AbstractInline
 {
+    /**
+     * {@inheritdoc}
+     */
+    public function isContainer(): bool
+    {
+        return true;
+    }
 }