+++ src/InlineParserEngine.php
@@ -14,11 +14,15 @@
 
 namespace League\CommonMark;
 
+use League\CommonMark\Block\Element\AbstractStringContainerBlock;
 use League\CommonMark\Inline\AdjacentTextMerger;
 use League\CommonMark\Inline\Element\Text;
 use League\CommonMark\Node\Node;
 use League\CommonMark\Reference\ReferenceMap;
 
+/**
+ * @internal
+ */
 final class InlineParserEngine
 {
     protected $environment;
@@ -29,10 +33,10 @@ final class InlineParserEngine
     }
 
     /**
-     * @param Node         $container
-     * @param ReferenceMap $referenceMap
+     * @param AbstractStringContainerBlock $container
+     * @param ReferenceMap                 $referenceMap
      */
-    public function parse(Node $container, ReferenceMap $referenceMap)
+    public function parse(AbstractStringContainerBlock $container, ReferenceMap $referenceMap)
     {
         $inlineParserContext = new InlineParserContext($container, $referenceMap);
         while (($character = $inlineParserContext->getCursor()->getCharacter()) !== null) {