+++ src/UnmatchedBlockCloser.php
@@ -16,6 +16,9 @@ namespace League\CommonMark;
 
 use League\CommonMark\Block\Element\AbstractBlock;
 
+/**
+ * @internal
+ */
 class UnmatchedBlockCloser
 {
     /**
@@ -65,6 +68,10 @@ class UnmatchedBlockCloser
 
     public function resetTip()
     {
+        if ($this->context->getTip() === null) {
+            throw new \RuntimeException('No tip to reset to');
+        }
+
         $this->oldTip = $this->context->getTip();
     }