+++ tests/unit/Block/Element/AbstractBlockTest.php
@@ -15,6 +15,7 @@
 namespace League\CommonMark\Tests\Unit\Block\Element;
 
 use League\CommonMark\Block\Element\AbstractBlock;
+use League\CommonMark\ContextInterface;
 use League\CommonMark\Inline\Element\AbstractInline;
 use PHPUnit\Framework\TestCase;
 
@@ -57,6 +58,26 @@ class AbstractBlockTest extends TestCase
 
         $block->setEndLine(42);
         $this->assertEquals(42, $block->getEndLine());
+    }
+
+    public function testFinalize()
+    {
+        $block = $this->getMockForAbstractClass(AbstractBlock::class);
+
+        $this->assertTrue($block->isOpen());
+
+        $context = $this->getMockForAbstractClass(ContextInterface::class);
+        $context->method('getTip')->willReturn($this->getMockForAbstractClass(AbstractBlock::class));
+        $block->finalize($context, 7);
+
+        $this->assertFalse($block->isOpen());
+        $this->assertEquals(7, $block->getEndLine());
+
+        // Try to re-finalize - there should be no effect
+        $block->finalize($context, 42);
+
+        $this->assertFalse($block->isOpen());
+        $this->assertEquals(7, $block->getEndLine());
     }
 
     private static function getMethod(string $name): \ReflectionMethod